Added passing Webhook send event through Redis Pub/Sub to handle horizontal scale issues
If a user connects to the WebSocket server of one service replica, it is likely that another replica will process the user's conversion. As a result, the first replica won’t be able to notify the user when the conversion is finished, since it isn’t connected to the user via WebSocket. To handle this, the processing replica sends an event to all other replicas. If one of them has an active WebSocket connection with the user, it forwards the message from the event to the user.